requirements.txt
numpy
seaborn
pandas
matplotlib
Pillow
tensorflow
scikit-learn
mlflow